+++ packages/inline-chat-kit/src/__tests__/public-api.test.ts
@@ -9,6 +9,7 @@ import * as kit from "../index";
  */
 const EXPECTED_EXPORTS = [
   "AddCardsOverlay",
+  "AnswerActions",
   "Button",
   "ChatHeader",
   "ChatInput",